# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "MyMechanicShop API"
  description: "API for a mechanic shop with endpoints for customers, mechanic staff, inventory, and service tickets"
  version: '1.0.0'
host: "127.0.0.1:5000"
schemes: 
  - 'http'
consumes:
  - 'application/json'
produces:
  - 'application/json'
securityDefinitions:
  bearerAuthCustomer:
    type: apiKey
    name: Authorization
    in: header
  bearerAuthMechanic:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /serviceticket:
    post:
      tags:
        - Service Tickets
      summary: "Create a service ticket"
      description: "Create a service ticket. This is a customer-specific token authenticated route. Customer must be logged in to create a service ticket request. Limit of 20 service ticket creations per hour."
      security:
        - bearerAuthCustomer: []
      parameters:
        - in: "body"
          name: "body"
          description: "Service ticket information"
          required: true
          schema:
            $ref: "#/definitions/CreateServiceTicketPayload"
      responses:
        201:
          description: "Successfully created service ticket"
          schema:
            $ref: "#/definitions/CreateServiceTicketResponse"
          examples:
            application/json:
              "id": 1
              "VIN": "1234567"
              "service_date": "2025-07-22"
              "service_desc": "Oil change"
              "customer": {
                "email": "john@email.com",
                "name": "John Doe",
                "phone": "(123)456-7890"
              }
              "mechanics": []
              "items": []

    get:
      tags:
        - Service Tickets
      summary: "Retrieve all service tickets"
      description: "Retrieve all service tickets. No log in required (no passwords or salaries are returned). Data gets cached for 60 seconds."
      responses:
        200:
          description: "Successfully retrieved all service tickets"
          schema:
            $ref: "#/definitions/AllServiceTickets"
          examples:
            application/json:
              [
                {
                  "id": 1,
                  "VIN": "1234567",
                  "service_date": "2025-07-22",
                  "service_desc": "Oil change",
                  "customer": {
                    "email": "john@email.com",
                    "name": "John Doe",
                    "phone": "(123)456-7890"
                  },
                  "mechanics": [
                    {
                      "email": "mech@email.com",
                      "name": "Ron",
                      "phone": "(123)456-7892"
                    }
                  ],
                  "items": [
                    {
                      "item": {
                        "name": "item 1",
                        "price": 1.00
                      },
                      "quantity": 1
                    },
                    {
                      "item": {
                        "name": "item 2",
                        "price": 2.00
                      },
                      "quantity": 2
                    }
                  ],
                }
              ]
  /serviceticket/{ticket_id}/assign-mechanic:
    put:
      tags:
        - Service Tickets
      summary: "Assign current mechanic to service ticket"
      description: "Add the mechanic that is currently logged in to the service ticket (ID given in path). This is a mechanic-specific token authenticated route. Mechanic must be logged in to assign themselves to the service ticket. For assigning other mechanics, use the '/serviceticket/{ticket_id}/edit' route to edit."
      security:
        - bearerAuthMechanic: []
      parameters:
        - in: "path"
          name: "ticket_id"
          description: "Numeric ID to identify the service ticket that should be modified."
          required: true
          schema:
            type: "integer"
      responses:
        200:
          description: "Successfully added mechanic to service ticket"
          schema:
            $ref: "#/definitions/AssignMechanicToServiceTicket"
          examples:
            application/json:
              message: "Mechanic 1 added to Service Ticket #1"
  /serviceticket/{ticket_id}/remove-mechanic:
    put:
      tags:
        - Service Tickets
      summary: "Remove current mechanic from service ticket"
      description: "Remove the mechanic that is currently logged in from the service ticket (ID given in path). This is a mechanic-specific token authenticated route. Mechanic must be logged in to remove themselves from the service ticket. For removing other mechanics, use the '/serviceticket/{ticket_id}/edit' route to edit."
      security:
        - bearerAuthMechanic: []
      parameters:
        - in: "path"
          name: "ticket_id"
          description: "Numeric ID to identify the service ticket that should be modified."
          required: true
          schema:
            type: "integer"
      responses:
        200:
          description: "Successfully removed mechanic from the service ticket"
          schema:
            $ref: "#/definitions/RemoveMechanicFromServiceTicket"
          examples:
            application/json:
              message: "Mechanic successfully removed from Service Ticket #1"
  /serviceticket/{ticket_id}/edit:
    put:
      tags:
        - Service Tickets
      summary: "Add/remove mechanics from service ticket"
      description: "Add/removes mechanics from the service ticket (ID provided in path). The mechanic currently logged in may add/remove other mechanics by their ids passed in. Mechanic-specific token authenticated route. List of mechanic IDs to add/remove should be included in the request."
      security:
        - bearerAuthMechanic: []
      parameters:
        - in: "path"
          name: "ticket_id"
          description: "Numeric ID to identify which service ticket to get"
          required: true
          schema:
            type: "integer"
        - in: "body"
          name: "body"
          description: "Mechanic IDs to add/remove"
          required: false
          schema:
            $ref: "#/definitions/EditServiceTicketMechanicsPayload"
      responses:
        200:
          description: "Successfully added/removed the mechanic(s) listed from the service ticket."
          schema:
            $ref: "#/definitions/EditServiceTicketMechanicsResponse"
          examples:
            application/json:
              "id": 1
              "VIN": "1234567"
              "service_date": "2025-07-22"
              "service_desc": "Oil change"
              "customer": {
                "email": "john@email.com",
                "name": "John Doe",
                "phone": "(123)456-7890"
              }
              "mechanics": [
                {
                  "email": "mech@email.com",
                  "name": "Ron",
                  "phone": "(123)456-7892"
                }
              ]
              "items": [
                {
                  "item": {
                    "name": "item 1",
                    "price": 1.00
                  },
                  "quantity": 1
                },
                {
                  "item": {
                    "name": "item 2",
                    "price": 2.00
                  },
                  "quantity": 2
                }
              ]
  /serviceticket/add_items:
    put:
      tags:
        - Service Tickets
      summary: "Add items to service ticket"
      description: "Add items from the inventory to the service ticket (ID provided in the body). This is a mechanic-specific token authenticated route. Mechanic must be logged in AND already assigned to the service ticket to add items to the service ticket. Quantity of the items are stored as well. Items with quantity 0 or less will not get added to the service ticket."
      security:
        - bearerAuthMechanic: []
      parameters:
        - in: "body"
          name: "body"
          description: "Provide service ticket ID, which items to add and quantity of each item."
          required: true
          schema:
            $ref: "#/definitions/AddItemsToServiceTicketPayload"
      responses:
        200:
          description: "Successfully added items to the service ticket"
          schema:
            $ref: "#/definitions/AddItemsToServiceTicketResponse"
          examples:
            application/json:
              "id": 1
              "VIN": "1234567"
              "service_date": "2025-07-22"
              "service_desc": "Oil change"
              "customer": {
                "email": "john@email.com",
                "name": "John Doe",
                "phone": "(123)456-7890"
              }
              "mechanics": [
                {
                  "email": "mech@email.com",
                  "name": "Ron",
                  "phone": "(123)456-7892"
                }
              ]
              "items": [
                {
                  "item": {
                    "name": "item 1",
                    "price": 1.00
                  },
                  "quantity": 1
                },
                {
                  "item": {
                    "name": "item 2",
                    "price": 2.00
                  },
                  "quantity": 2
                }
              ]
definitions:
...
  CreateServiceTicketPayload:
    type: "object"
    properties:
      VIN:
        type: "string"
      service_date:
        type: "string"
        format: "date"
      service_desc:
        type: "string"
  CreateServiceTicketResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      VIN:
        type: "string"
      service_date:
        type: "string"
        format: date
      service_desc:
        type: "string"
      customer:
        type: "object"
        properties:
          name:
            type: "string"
          email:
            type: "string"
          phone:
            type: "string"
      mechanics:
        type: "array"
        items: 
          type: "object"
          properties:
            name:
              type: "string"
            email:
              type: "string"
            phone:
              type: "string"
      items:
        type: "array"
        items:
          type: "object"
          properties:
            item:
              type: "object"
              properties:
                name:
                  type: "string"
                price:
                  type: "number"
                  format: "float"
            quantity:
              type: "integer"
  AllServiceTickets:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        VIN:
          type: "string"
        service_date:
          type: "string"
          format: date
        service_desc:
          type: "string"
        customer:
          type: "object"
          properties:
            name:
              type: "string"
            email:
              type: "string"
            phone:
              type: "string"
        mechanics:
          type: "array"
          items: 
            type: "object"
            properties:
              name:
                type: "string"
              email:
                type: "string"
              phone:
                type: "string"
        items:
          type: "array"
          items:
            type: "object"
            properties:
              item:
                type: "object"
                properties:
                  name:
                    type: "string"
                  price:
                    type: "number"
                    format: "float"
              quantity:
                type: "integer"
  AssignMechanicToServiceTicket:
    type: "object"
    properties:
      message:
        type: "string"
  RemoveMechanicFromServiceTicket:
    type: "object"
    properties:
      message:
        type: "string"
  EditServiceTicketMechanicsPayload:
    type: "object"
    properties:
      add_mechanic_ids:
        type: "array"
        items:
          type: "integer"
        default: []
      remove_mechanic_ids:
        type: "array"
        items:
          type: "integer"
        default: []
  EditServiceTicketMechanicsResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      VIN:
        type: "string"
      service_date:
        type: "string"
        format: date
      service_desc:
        type: "string"
      customer:
        type: "object"
        properties:
          name:
            type: "string"
          email:
            type: "string"
          phone:
            type: "string"
      mechanics:
        type: "array"
        items: 
          type: "object"
          properties:
            name:
              type: "string"
            email:
              type: "string"
            phone:
              type: "string"
      items:
        type: "array"
        items:
          type: "object"
          properties:
            item:
              type: "object"
              properties:
                name:
                  type: "string"
                price:
                  type: "number"
                  format: "float"
            quantity:
              type: "integer"
  AddItemsToServiceTicketPayload:
    type: "object"
    properties:
      ticket_id:
        type: "integer"
      item_quant:
        type: "array"
        items:
          type: "object"
          properties:
            item_id:
              type: "integer"
            quantity:
              type: "integer"
  AddItemsToServiceTicketResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      VIN:
        type: "string"
      service_date:
        type: "string"
        format: date
      service_desc:
        type: "string"
      customer:
        type: "object"
        properties:
          name:
            type: "string"
          email:
            type: "string"
          phone:
            type: "string"
      mechanics:
        type: "array"
        items: 
          type: "object"
          properties:
            name:
              type: "string"
            email:
              type: "string"
            phone:
              type: "string"
      items:
        type: "array"
        items:
          type: "object"
          properties:
            item:
              type: "object"
              properties:
                name:
                  type: "string"
                price:
                  type: "number"
                  format: "float"
            quantity:
              type: "integer"